+++ android/app/src/main/AndroidManifest.xml
@@ -3,7 +3,8 @@
    <application
         android:label="AlertMe"
         android:name="${applicationName}"
-        android:icon="@mipmap/ic_launcher">
+        android:icon="@mipmap/ic_launcher"
+        android:enableOnBackInvokedCallback="true">
         <activity
             android:name=".MainActivity"
             android:exported="true"
