--- lib/pages/enter_otp_page.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-class EnterOtpPage extends StatelessWidget {
-  const EnterOtpPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return SafeArea(
-      child: Scaffold(
-        body: Form(
-          child: ListView(),
-        ),
-      ),
-    );
-  }
-}
